moved delete_all_files to admin screen
