+++ lib/camera_api/admin/delete_all_button.dart
