+++ lib/camera_api/admin/reset_button.dart
